add class login and register
