+++ src/classes/listeners/class.listener-register.js
@@ -1,6 +1,7 @@
 export class RegisterListener {
     constructor(kanban) {
         this.kanban = kanban;
+        this.register = kanban.register;
         this.activateListener();
     }
 
@@ -8,9 +9,9 @@ export class RegisterListener {
         const passwordField = document.getElementById(inputFieldId);
         if (!passwordField) return;
 
-        const togglePasswordVisibility = (event) => togglePassword(event, passwordField);
-        const updateBackgroundOnFocus = () => updateBackgroundImage(passwordField, false);
-        const resetOnBlur = () => resetState(passwordField);
+        const togglePasswordVisibility = (event) => this.register.togglePassword(event, passwordField);
+        const updateBackgroundOnFocus = () => this.register.updateBackgroundImage(passwordField, false);
+        const resetOnBlur = () => this.register.resetState(passwordField);
 
         passwordField.addEventListener('click', togglePasswordVisibility);
         passwordField.addEventListener('focus', updateBackgroundOnFocus);
@@ -32,25 +33,25 @@ export class RegisterListener {
 
 
     activateListener() {
-        document.getElementById('signupForm')?.addEventListener('submit', submitData);
-        document.getElementById('privacy-policy')?.addEventListener('click', checkBoxClicked);
-        document.getElementById('back-btn')?.addEventListener('click', forwardToIndex);
-        document.getElementById('privacy-policy-link')?.addEventListener('click', forwardPrivacy);
-        document.getElementById('legal-notice-link')?.addEventListener('click', forwardLegal);
-        addPasswordFieldToggle('password');
-        addPasswordFieldToggle('confirmPassword');
+        document.getElementById('signupForm')?.addEventListener('submit', this.register.submitData);
+        document.getElementById('privacy-policy')?.addEventListener('click', this.register.checkBoxClicked);
+        document.getElementById('back-btn')?.addEventListener('click', this.kanban.forwardToIndex);
+        document.getElementById('privacy-policy-link')?.addEventListener('click', this.kanban.forwardPrivacy);
+        document.getElementById('legal-notice-link')?.addEventListener('click', this.kanban.forwardLegal);
+        this.addPasswordFieldToggle('password');
+        this.addPasswordFieldToggle('confirmPassword');
     }
 
 
 
     deactivateAllListenersRegister() {
-        document.getElementById('signupForm')?.removeEventListener('submit', submitData);
-        document.getElementById('privacy-policy')?.removeEventListener('click', checkBoxClicked);
-        document.getElementById('back-btn')?.removeEventListener('click', forwardToIndex);
-        document.getElementById('privacy-policy-link')?.removeEventListener('click', forwardPrivacy);
-        document.getElementById('legal-notice-link')?.removeEventListener('click', forwardLegal);
-        removePasswordFieldToggle('password');
-        removePasswordFieldToggle('confirmPassword');
+        document.getElementById('signupForm')?.removeEventListener('submit', this.register.submitData);
+        document.getElementById('privacy-policy')?.removeEventListener('click', this.register.checkBoxClicked);
+        document.getElementById('back-btn')?.removeEventListener('click', this.kanban.forwardToIndex);
+        document.getElementById('privacy-policy-link')?.removeEventListener('click', this.kanban.forwardPrivacy);
+        document.getElementById('legal-notice-link')?.removeEventListener('click', this.kanban.forwardLegal);
+        this.removePasswordFieldToggle('password');
+        this.removePasswordFieldToggle('confirmPassword');
     }
 
 }
